+++ service/controller/machine_deployment.go
@@ -72,6 +72,7 @@ type MachineDeploymentConfig struct {
 	Logger             micrologger.Logger
 	RandomKeysSearcher randomkeys.Interface
 
+	AlikeInstances             string
 	CalicoCIDR                 int
 	CalicoMTU                  int
 	CalicoSubnet               string
@@ -549,6 +550,7 @@ func newMachineDeploymentResources(config MachineDeploymentConfig) ([]resource.I
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 
+			AlikeInstances:   config.AlikeInstances,
 			InstallationName: config.InstallationName,
 		}